backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
PyJWT==2.9.0
httpx==0.27.0
supabase==2.6.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
